Remove "splitter" backend
We support different storage backends (particularly, Django's ORM and Elastic
Search). We had a method ("splitter") to write data to both, read from the
ORM, and search using Elastic. That's kind of silly, though -- if you are
using the ORM and Elastic, you can just as easily be using haystack.