Rename field, add doc_type to layer model
Taking another pass at implementing layers for arbitrary data types, this adds
a doc_type field to the Layer model, in addition to renaming "reference" to
"doc_id". This also modifies the data present so that the previous delimiter
(:) is now a forward slash (convenient to match URLs, as we'll see soon).