Split dependencies of backends during tests.
We've been testing with both the haystack and elastic dependencies installed
at once. This will be a problem when working with the postgres backend because
it requires a specific version of Django. We'd like to keep testing the other
versions of Django with haystack and elastic, so we need to set up a framework
for testing separate backends with different dependencies installed.