Store "doc_root" redundantly for ease of querying.
When users search, they provide a document id ("regulation") and a version. We
had not been able to filter by thus document id as that data's not a separate
field. While it may make more sense to store this with the documents, in the
short term, we can store it in the search index.