Account for more corner cases w/ section search.
Case 1: A search matches a node's title, but the node doesn't have text. In
this situation, we return text for the first child of that node.

Case 2: A search which matches a section, but none of its paragraphs contain
all of the search term. Here, we return text for the first child of the
section.